Update the AP section, talk about ROC and AUC and Precision recall curves
The amendment comments: The acronym for Average Precision (AP) should come with its introduction. The section also needs a short description of the receiving operating characteristic (ROC) curve. Thresholding is mentioned in the section but not explained. A short explanation comes later (precision-recall curve) but that section should be placed earlier and corroborated with a picture and a better explanation on its reliance on classification confidence scores.